search zbs

+++ WebApplication1/WebApplication1/Models/AccountViewModels.cs
@@ -8,7 +8,7 @@ namespace WebApplication1.Models
         
         public AspNetUser User { get; set; }
         public ICollection<Demotivator> Demotivator { get; set; }
-        public Single Rate { get; set; }
+        public Double Rate { get; set; }
     }
